Move the Kubernetes models
Move the models used by the Kubernetes storage layer out of the
gafaelfawr.storage.kubernetes module and into their own module.
Add a parent class for arbitrary Kubernetes resources, and rename
the status class since it's also generic, in preparation for
managing additional types of resources in the future.
